Move Icons into its own component file
