+++ app/src/features/ApplicationWorkspace/constants.ts
@@ -1,13 +1,3 @@
-import {
-  Check,
-  Email,
-  ExpandMore,
-  LocalPhone,
-  RadioButtonChecked,
-  ShortText,
-  Subject,
-} from '@mui/icons-material';
-
 export type FieldTypes =
   | 'short_text'
   | 'long_text'
@@ -150,40 +140,4 @@ export const fields = {
       type: 'yes_no',
     },
   ],
-};
-
-interface FieldTypeOption {
-  icon: React.ReactNode;
-  color: string;
-}
-
-export const FieldTypeOptions: Record<FieldTypes, FieldTypeOption> = {
-  short_text: {
-    icon: <ShortText />,
-    color: 'primary.main',
-  },
-  long_text: {
-    icon: <Subject />,
-    color: 'secondary.main',
-  },
-  number: {
-    icon: <LocalPhone />,
-    color: 'warning.main',
-  },
-  yes_no: {
-    icon: <RadioButtonChecked />,
-    color: 'success.main',
-  },
-  multiple_choice: {
-    icon: <Check />,
-    color: 'success.main',
-  },
-  email: {
-    icon: <Email />,
-    color: 'error.main',
-  },
-  dropdown: {
-    icon: <ExpandMore />,
-    color: 'info.main',
-  },
 };
